Refactoring `org.bdgenomics.adam.io` package.
* Added documentation to all classes, methods, and fields that were missing
  descriptions.
* Factored duplicated code from `SingleFastqRecordReader` and
  `InterleavedFastqRecordReader` out into a new abstract base class
  `FastqRecordReader`.
* Tightened up access modifiers where possible. Sadly, the two input formats
  cannot be made more private, since Java has a more restricted form of
  package-private than scala, and the input formats are used from inside of
  `org.bdgenomics.adam.rdd`.
* Removed `@author` tags.
